update timers in Y
RecvCircuit is a one-off stage of the protocol,
which was taking a significant amount of time.
This commit resets timer after this stage,
so that timer only measures repeated stages of the protocol.

OT stages are also timed; these stages are usually fast
and don't make for a large fraction of each round's time.